Create Manga Rating Repository and Service with find by User and Manga methods, change manga rating script and Rating property type to int in Manga Rating class

+++ src/main/java/com/NowakArtur97/WorldOfManga/controller/mangaRating/MangaRatingController.java
@@ -22,7 +22,7 @@ public class MangaRatingController {
 
 	@GetMapping(path = "/rateManga")
 	public String rateManga(HttpServletRequest request, @RequestParam("id") Long mangaId,
-			@RequestParam("rating") Long rating) throws MangaNotFoundException {
+			@RequestParam("rating") int rating) throws MangaNotFoundException {
 
 		userService.rateManga(mangaId, rating);
